Rename TropoDB table files
