+++ implementation/rocksdb/db/zns_impl/impl/tropodb_impl_background.cc
@@ -24,8 +24,8 @@
 #include "db/zns_impl/persistence/tropodb_manifest.h"
 #include "db/zns_impl/persistence/tropodb_wal.h"
 #include "db/zns_impl/persistence/tropodb_wal_manager.h"
-#include "db/zns_impl/table/zns_sstable_manager.h"
-#include "db/zns_impl/table/zns_table_cache.h"
+#include "db/zns_impl/table/tropodb_sstable_manager.h"
+#include "db/zns_impl/table/tropodb_table_cache.h"
 #include "db/zns_impl/utils/tropodb_logger.h"
 #include "port/port.h"
 #include "rocksdb/db.h"
