+++ implementation/rocksdb/db/zns_impl/index/tropodb_version_edit.cc
@@ -5,7 +5,7 @@
 
 #include "db/zns_impl/tropodb_config.h"
 #include "db/zns_impl/index/tropodb_version.h"
-#include "db/zns_impl/table/zns_zonemetadata.h"
+#include "db/zns_impl/table/tropodb_zonemetadata.h"
 #include "db/zns_impl/utils/tropodb_logger.h"
 #include "rocksdb/rocksdb_namespace.h"
 #include "util/coding.h"
