+++ implementation/rocksdb/db/zns_impl/table/iterators/sstable_ln_iterator.h
@@ -11,8 +11,8 @@
 
 #include "db/dbformat.h"
 #include "db/zns_impl/table/iterators/iterator_wrapper.h"
-#include "db/zns_impl/table/zns_sstable_manager.h"
-#include "db/zns_impl/table/zns_zonemetadata.h"
+#include "db/zns_impl/table/tropodb_sstable_manager.h"
+#include "db/zns_impl/table/tropodb_zonemetadata.h"
 #include "rocksdb/iterator.h"
 #include "rocksdb/slice.h"
 #include "rocksdb/status.h"
