+++ implementation/rocksdb/db/zns_impl/table/tropodb_sstable_builder.cc
@@ -1,7 +1,7 @@
-#include "db/zns_impl/table/zns_sstable_builder.h"
+#include "db/zns_impl/table/tropodb_sstable_builder.h"
 
 #include "db/zns_impl/tropodb_config.h"
-#include "db/zns_impl/table/ln_zns_sstable.h"
+#include "db/zns_impl/table/tropodb_ln_sstable.h"
 #include "db/zns_impl/utils/tropodb_logger.h"
 
 namespace ROCKSDB_NAMESPACE {
