+++ implementation/rocksdb/db/zns_impl/table/tropodb_sstable_builder.h
@@ -3,8 +3,8 @@
 #ifndef ZNS_SSTABLE_BUILDER_H
 #define ZNS_SSTABLE_BUILDER_H
 
-#include "db/zns_impl/table/zns_sstable.h"
-#include "db/zns_impl/table/zns_zonemetadata.h"
+#include "db/zns_impl/table/tropodb_sstable.h"
+#include "db/zns_impl/table/tropodb_zonemetadata.h"
 #include "rocksdb/rocksdb_namespace.h"
 
 namespace ROCKSDB_NAMESPACE {
