+++ implementation/rocksdb/db/zns_impl/table/tropodb_sstable_manager.h
@@ -10,10 +10,10 @@
 #include "db/zns_impl/io/szd_port.h"
 #include "db/zns_impl/memtable/tropodb_memtable.h"
 #include "db/zns_impl/ref_counter.h"
-#include "db/zns_impl/table/l0_zns_sstable.h"
-#include "db/zns_impl/table/ln_zns_sstable.h"
-#include "db/zns_impl/table/zns_sstable.h"
-#include "db/zns_impl/table/zns_zonemetadata.h"
+#include "db/zns_impl/table/tropodb_l0_sstable.h"
+#include "db/zns_impl/table/tropodb_ln_sstable.h"
+#include "db/zns_impl/table/tropodb_sstable.h"
+#include "db/zns_impl/table/tropodb_zonemetadata.h"
 #include "rocksdb/slice.h"
 #include "rocksdb/status.h"
