+++ implementation/rocksdb/db/zns_impl/tests/zns_sstable_manager_test.cc
@@ -1,5 +1,5 @@
 
-#include "db/zns_impl/table/zns_sstable_manager.h"
+#include "db/zns_impl/table/tropodb_sstable_manager.h"
 
 #include "db/zns_impl/tests/zns_test_utils.h"
 #include "test_util/testharness.h"
