+++ implementation/rocksdb/db/zns_impl/tropodb_impl.h
@@ -32,8 +32,8 @@
 #include "db/zns_impl/persistence/tropodb_manifest.h"
 #include "db/zns_impl/persistence/tropodb_wal.h"
 #include "db/zns_impl/persistence/tropodb_wal_manager.h"
-#include "db/zns_impl/table/zns_sstable_manager.h"
-#include "db/zns_impl/table/zns_zonemetadata.h"
+#include "db/zns_impl/table/tropodb_sstable_manager.h"
+#include "db/zns_impl/table/tropodb_zonemetadata.h"
 #include "options/cf_options.h"
 #include "port/port.h"
 #include "rocksdb/db.h"
